+++ libraries/DeanLib/include/PerformanceTracker.h
@@ -15,7 +15,7 @@ public:
 
 	void startTracking(const std::string& trackerName);
 	void stopTracking(const std::string& trackerName);
-	double getElapsedTime(const std::string& trackerName);
+	float getElapsedTime(const std::string& trackerName);
 	void removeTracker(const std::string& trackerName);
 	void clearTracker(const std::string& trackerName);
